Move select location components to new directory and split components up
